requirements.txt
Flask
Pillow
pytesseract
opencv-python-headless
pdf2image
transformers
torch